+++ SleepHunter/Models/Inventory.cs
@@ -24,6 +24,8 @@ namespace SleepHunter.Models
 
         public int Count => inventory.Count((item) => { return !item.IsEmpty; });
 
+        public IEnumerable<InventoryItem> AllItems => inventory;
+
         public IEnumerable<string> ItemNames => 
             from i in inventory where !i.IsEmpty && !string.IsNullOrWhiteSpace(i.Name) select i.Name;
 
@@ -101,6 +103,7 @@ namespace SleepHunter.Models
             }
 
             reader.BaseStream.Position = inventoryPointer;
+            
 
             for (int i = 0; i < inventoryVariable.Count; i++)
             {